This is a re-implementation of xmlrpclib's xml -> python object
This is a re-implementation of xmlrpclib's xml -> python object
unmarshaller in C.  Using this instead of the python version shaves
about 40% off of the time it takes to unmarshalling large xml-rpc
responses (i.e. querying for all of the packages from the RH73
channel).

Unfortunately, the query is still way too slow.  And the way xmlrpclib
uses expat for XML parsing breaks the profiler
(https://sourceforge.net/tracker/?func=detail&atid=105470&aid=534864&group_id=5470)
in the version of python that I got with RH73.  And I really don't
feel like building some bleeding-edge python from a tarball right now,
so I'm not sure where the remaining bottlenecks are.

svn path=/trunk/; revision=19