Profilers?  Heh!  Profilers are for the weak!
It turns out that we can cut the processing time of big xml-rpc responses
by over 50% just by reading the whole response in one giant slurp instead
of reading it in 1024-byte chunks.

This is almost certainly a bad thing to do in the non-local case, so we'll
ultimately want to figure out if this is a local or remote read and
Do The Right Thing(tm).

svn path=/trunk/; revision=21