Merged my deal-with-a-dead-server code w/ Joe's deal-with-a-dead-server
Merged my deal-with-a-dead-server code w/ Joe's deal-with-a-dead-server
code.  We now ping the server while the "we are dead" dialog is up,
and remove it (and redo the failed xmlrpc call) if the server comes back
to life.

Made the ChannelOption widget a ServerListener, so it will change if the
channel line-up changes.

ServerListener-related fixes.  Removed some race conditions.

Channel caching fixes.

New bugs introduced.

svn path=/trunk/; revision=81