Hacking: Copied the faults over from rc's rcfault into rcd_util.fault.
Hacking:
* Copied the faults over from rc's rcfault into rcd_util.fault.

* Made the throbber take a height and width which we get from the toolbar
and use them.

* Added "Install from File..." and "Install from URL..." menu items under
the "File" menu and implemented them.  The former pops up a file selection
dialog and allows the user to select files.  It warns the user if any
package is invalid or the list is empty.  Then it just passes them onto
the daemon and adds them to the pending actions list.  The latter pops up
a small window asking for the URL, and then proceeds to download the file
in a separate thread.  Once the file is downloaded, it's sent to the
daemon and (if it's valid) gets added to the pending actions list.

svn path=/trunk/; revision=99