More hacking:
* Make the client always connect locally to the daemon on startup.  We'll
need to add some --host, --user, --password command-line flags for changing
this.  Separate the UI and the rcd_util bits.

* Added a "Connect..." item to the "File" menu so we can change the server
after we're running.

* Get rid of the old History view and make Tambet's the only one.

* Start on doing mount/unmount but become daunted at the bastard that is
GtkFileSelection and its seeming inability to select directories,
remember that it has that annoying parent directory behavior, and decide
to give up for the day and work on it tomorrow.

svn path=/trunk/; revision=104